Test: Offer different resolution and generalization methods for:
- old resolution - generalization mapping without 5 & 150
  (for data that has not been updated, yet)
- new resolution - generalization mapping based on
  https://confluence.geops.ch/pages/viewpage.action?spaceKey=TRAFIMAGETEAM&title=Gruppierung+Vector+Tiles
